frontend to backend communication with websockets, now starting JS connection
