+++ srcs/backend/pong/GameObject.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+@dataclass
+class GameObject:
+    x: float
+    y: float
+    width: float
+    height: float
