+++ srcs/backend/pong/commands.txt
@@ -1,4 +1,4 @@
 docker build -t pong .
 docker rm -f $(docker ps -qa)
-docker run -p 8080:8080 pong
+docker run -p 8888:8888 pong
