+++ srcs/frontend/conf/nginx.conf
@@ -51,5 +51,15 @@ http {
         location /app/ {
             proxy_pass http://transcendence:8000/app/;
         }
+	
+		location /ws/pong/ {
+			proxy_pass http://pong:8080/;
+			proxy_set_header Host $host;
+			proxy_set_header X-Real-IP $remote_addr;
+			proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+			proxy_set_header X-Forwarded-Proto https;
+			proxy_set_header Upgrade $http_upgrade;
+			proxy_set_header Connection "upgrade";
+        }
     }
 }
