+++ srcs/frontend/requirements.txt
@@ -1,0 +1,4 @@
+websockets
+eventlet
+requests
+python-socketio==5.0.4
